Fix all changes flake8 pointed out and move code to map in experiments
